+++ packages/core/src/utils/providers.spec.ts
@@ -1,3 +1,4 @@
+import 'reflect-metadata';
 import { Logger, LoggerConfig } from '../types/logger';
 import { ServiceProvider } from '../types/mix';